feat: move dragdropfile & fix test error
