--- frontend/src/components/DragDropFile/index.ts
@@ -1,1 +1,0 @@
-export { DragDropFile } from "./DragDropFile";
