--- frontend/src/components/DragDropFile/mockdata.json
@@ -1,7 +1,0 @@
-{
-  "file": {
-    "content": "Test File DragnDrop",
-    "name": "test.bpmn",
-    "type": "text/plain"
-  }
-}
